pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smdgf"
version = "0.1.0"
description = "Social mind data generation framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.12,<3",
  "typer>=0.15,<1",
  "hydra-core>=1.3,<1.4",
  "pyyaml>=6,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "ruff>=0.12,<0.13",
]

[project.scripts]
smdgf = "smdgf.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
